refactor: change api func names
