+++ api/index.ts
@@ -1,1 +1,3 @@
-export { useApi } from './client/use-api';
+export { useApi } from './use-api';
+export { GetAllUsersClient } from './users/users-client';
+export { GetAllUsersServer } from './users/users-server';
