+++ api/users/users-server.ts
@@ -2,7 +2,7 @@
 
 export const revalidate = 60;
 
-export async function fetchUsersSSR() {
+export async function GetAllUsersServer() {
   try {
     const res = await fetch(
       `${process.env.NEXT_PUBLIC_BACK}/users/public?page=1&limit=5`,
