+++ app/users/page.tsx
@@ -1,12 +1,12 @@
 // app/users/page.tsx
 
-import { fetchUsersSSR } from '@/api/server/users';
+import { GetAllUsersServer } from '@/api/users/users-server';
 import { UsersFilter, UsersList } from './components';
 
 export const revalidate = 60;
 
 export default async function UsersPage() {
-  const data = await fetchUsersSSR();
+  const data = await GetAllUsersServer();
 
   return (
     <section className="flex flex-col gap-8 py-8 md:py-10 m-auto">
